requirements.txt
twilio==9.0.0
python-dotenv==1.0.1
pywhatkit==5.4
